+++ CHANGES.rst
@@ -4,8 +4,10 @@ Changelog
 3.12 (unreleased)
 -----------------
 
-- Nothing changed yet.
-
+- Added parameter `bypass_check_permission` to method
+  `DocumentGenerationHelperView.display` to bypass read permission check,
+  this is useful when context is an object that was get unrestrictively.
+  [gbastien]
 
 3.11 (2019-12-03)
 -----------------